Add a workaround for using autochdir
Related to justinmk#19

Allow the :cd on autocmd emulation of autochdir to work with dirvish.
Describe this workaround for autochdir-like behavior in the help.

This change might be a first step towards autochdir, but it's not enough
yet.

The autocmd breaks dirvish when we try to force an alternate file
because we're :cd-ing while dirvish is trying to do stuff and we end up
with an empty buffername. This change tracks when we're in set_altbuf so
the autocmd can be disabled during that region (using noau didn't work
and replacing set_altbuf with keepalt everywhere didn't work either).

Ideally, we detect if autochdir is enabled and re-enable it when we're
clear of set_altbuf, but it doesn't work yet.