Add option for preamble before shdo scripts
I want all of my scripts to start in the current dirvish working
directory, so I use the example from the doc:
  let g:dirvish_shdo_preamble = 'cd {}'

Other users may tend to create scripts with arguments, want their
commands to be inside a function, or output some diagnostic data.

Using the same escaping that is applied on lines since that seems like
the safest thing to do.